packages/engine/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pandas==2.1.4
numpy==1.26.4
pandas-ta==0.3.14b1
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.4
aiohttp==3.11.11
